Add basic beginning structure to html
Added the begginings of the structure to the html code based off of the design template and psuedo code made
for this project. This will likely undergo several changes and iterations before final submission, so much
of the labeling, div classifacation and input methods are deliberately left ambiguous and blank.