Add min-width to splash image and set background image settings
Removed the img tag in html and used css to import the background image. I set a min-width of 400px and
set it's width to 40% viewport width and 100vh just to get a feeling for how this will look closer to
completetion. Undoubtedly many of these settings will change after styling the form