Add paragraph element for invalid password
While the element has been made with basic styling, a simple JS source needs to
still be made to implement change values between visible and hidden. This is also the case for user_password input. The item should
remain uncolored until either a)the passwords match and meet minimum requirements or b)are invalid or left blank.