Adds password eye to the password fields in form
this was done with the aid of a tutorial from geeksforgeeks
https://www.geeksforgeeks.org/how-to-toggle-password-visibility-using-html-and-javascript/
The images used are from their website and temporary. The reason for both password fields being alternated between visible and none
visible is for user experience. They can choose to show their passwords by clicking in either field and compare them. Allowing the user
to see where they blundered in writing out their password.