Fix password styling staying green when passwords don't match or regex fails
This commit adds several items. First it takes both confirm and password selectors and creates
a keyup event listener to execute the 'validate' class add code. It also sets any error message variables to hidden should the user
meet the criteria for both match passwords and regex calls. Secondly, a new function and event listener are added to both the html
and js. This function acts as a double check should the user delete or accidently add an additonal character when moving away
from the input box. The reCheck button is called when the fields are out of focus and if it passes this function (fails the tests)
the form class validated is deleted and the user cannot submit the form.